+++ docker/sds-api/spine-directory-service/sds/request/cpm.py
@@ -5,6 +5,8 @@ from typing import List
 from lookup.sds_exception import SDSException
 from request.base_handler import ORG_CODE_QUERY_PARAMETER_NAME, ORG_CODE_FHIR_IDENTIFIER, \
     IDENTIFIER_QUERY_PARAMETER_NAME, SERVICE_ID_FHIR_IDENTIFIER, PARTY_KEY_FHIR_IDENTIFIER
+from request.cpm_config import DEVICE_FILTER_MAP, ENDPOINT_FILTER_MAP, DEVICE_DATA_MAP, ENDPOINT_DATA_MAP, DEFAULT_ENDPOINT_DICT, DEFAULT_DEVICE_DICT
+
 
 async def get_device_from_cpm(ods_code: str, interaction_id: str, manufacturing_organization: str = None, party_key: str = None) -> List:
     return [
@@ -16,6 +18,7 @@ async def get_device_from_cpm(ods_code: str, interaction_id: str, manufacturing_
         }
     ]
 
+
 async def get_endpoint_from_cpm(ods_code: str, interaction_id: str = None, party_key: str = None) -> List:
     return [
         {
@@ -25,99 +28,114 @@ async def get_endpoint_from_cpm(ods_code: str, interaction_id: str = None, party
         }
     ]
 
-def filter_cpm_devices_response(data: dict, query_parts: dict):
-    if "org_code" not in query_parts or "interaction_id" not in query_parts or not query_parts["org_code"] or not query_parts["interaction_id"]:
-            raise SDSException("org_code and interaction_id must be provided")
+def process_cpm_endpoint_request(data: dict, query_parts: dict):
+    endpoints = EndpointCpm(data=data, query_parts=query_parts)
+    filtered_endpoints = endpoints.filter_cpm_response()
+    return endpoints.transform_to_ldap(filtered_endpoints)
+
+def process_cpm_device_request(data: dict, query_parts: dict):
+    devices = DeviceCpm(data=data, query_parts=query_parts)
+    filtered_devices = devices.filter_cpm_response()
+    return devices.transform_to_ldap(filtered_devices)
+
+class BaseCpm:
+    FILTER_MAP = {}
+    DATA_MAP = {}
+    DEFAULT_DICT = {}
     
-    query_mapping = {
-        "org_code": "Owner",
-        "interaction_id": "InteractionIds",
-        "manufacturing_organization": "ManufacturingOdsCode",
-        "party_key": "PartyKey"
-    }
-    return filter_cpm_response(data, query_parts, query_mapping)
+    def __init__(self, data, query_parts):
+        self.data = data
+        self.query_parts = query_parts
 
-def filter_cpm_endpoints_response(data: dict, query_parts: dict):
-    non_empty_count = sum(1 for value in query_parts.values() if value and value != 0)
-    if non_empty_count < 2:
-        _raise_invalid_query_params_error()
-
-    query_mapping = {
-        "org_code": "managingOrganization",
-        "service_id": "interactionID",
-        "party_key": "mhsPartyKey"
-    }
-    return filter_cpm_response(data, query_parts, query_mapping)
-
-def filter_cpm_response(data: dict, query_parts: dict, query_mapping: dict):
-    filtered_results = []
-    filters = {key: False for key in query_parts}
+    def filter_cpm_response(self):
+        filtered_results = []
+        filters = {key: False for key in self.query_parts}
+        
+        for result in self.data["entry"]:
+            for index, res in enumerate(result["entry"]) if "resourceType" in result and result["resourceType"] == "Bundle" else []:
+                for service in res["item"] if "resourceType" in res and res["resourceType"] == "QuestionnaireResponse" else []:
+                    filters = self._check_each_item(self, filters, service)
+                
+            all_filters_true = all(value for value in filters.values())
+            if all_filters_true:
+                filtered_results.append(result["entry"])
+            filters = {key: False for key in filters}
+        
+        return filtered_results
     
-    for result in data["entry"]:
-        if "resourceType" in result and result["resourceType"] == "Bundle":
-            for index, res in enumerate(result["entry"]):
-                if "resourceType" in res and res["resourceType"] == "QuestionnaireResponse":
-                    for service in res["item"]:
-                        for key, value in query_parts.items():
-                            if service["text"] == query_mapping[key]:
-                                for answer in service["answer"]:
-                                    if answer["valueString"] == value:
-                                        filters[key] = True                   
-        all_filters_true = all(value for value in filters.values())
-        if all_filters_true:
-            filtered_results.append(result["entry"])
-        filters = {key: False for key in filters}
+    @staticmethod
+    def _check_each_item(self, filters, service):
+        for key, value in self.query_parts.items():
+            if service["text"] == self.FILTER_MAP[key]:
+                filters[key] = self._check_match(self, service["answer"], value)
+        return filters
     
-    return filtered_results
+    @staticmethod
+    def _check_match(self, answers, match):
+        for answer in answers:
+            if answer["valueString"] == match:
+                return True
+    
+    @staticmethod
+    def process_device_response(self, item):
+        if "resourceType" in item and item["resourceType"] == "Device":
+            if "identifier" in item:
+                for identifier in item["identifier"]:
+                    if identifier.get("system") == "https://fhir.nhs.uk/Id/nhsSpineASID":
+                        return identifier.get("value")
+        return ""
+    
+    @staticmethod
+    def process_questionnaire_response(self, item, data_dict, ldap_data_mapping):
+        if "resourceType" in item and item["resourceType"] == "QuestionnaireResponse":      
+            for service in item.get("item", []):
+                if service["text"] in ldap_data_mapping:
+                    for answer in service["answer"]:
+                        key = ldap_data_mapping[service["text"]]
+                        value = answer.get("valueInteger", answer.get("valueString"))
+                        data_dict = self.get_data(data_dict, key, value)
+                            
+        return data_dict
+    
+    @staticmethod
+    def get_data(data_dict, key, value):
+        if isinstance(data_dict[key], list):
+            data_dict[key].append(value)
+        elif isinstance(data_dict[key], int):
+            data_dict[key] = value
+        else:
+            data_dict[key] = value
+        
+        return data_dict
+    
+    def transform_to_ldap(self, data: List) -> List:
+        ldap_data = []
+        
+        for d in data:
+            data_dict = copy.deepcopy(self.DEFAULT_DICT)
+            for item in d:
+                data_dict = self.process_questionnaire_response(self, item, data_dict, self.DATA_MAP)
 
-def transform_device_to_SDS(data: List) -> List:
-    ldap_data = []
-    default_data_dict = dict(
-        nhsAsClient = [],
-        nhsAsSvcIA = [],
-        nhsMhsManufacturerOrg = "",
-        nhsMhsPartyKey = "",
-        nhsIdCode = "",
-        uniqueIdentifier = ""
-    )
+            ldap_data.append(data_dict)
 
-    for d in data:
-        data_dict = copy.deepcopy(default_data_dict)
-        for item in d:
-            if "resource" in item:
-                data_dict["uniqueIdentifier"] = process_device_resource(item["resource"])
-            else:
-                data_dict = process_questionnaire_response(item, data_dict)
+        return ldap_data
+    
 
-        ldap_data.append(data_dict)
-
-    return ldap_data
-
-
-def process_device_resource(item):
-    if "resourceType" in item and item["resourceType"] == "Device":
-        if "identifier" in item:
-            for identifier in item["identifier"]:
-                if identifier.get("system") == "https://fhir.nhs.uk/Id/nhsSpineASID":
-                    return identifier.get("value")
-    return ""
-
-def process_questionnaire_response(item, data_dict):
-    if "resourceType" in item and item["resourceType"] == "QuestionnaireResponse":
-        for service in item.get("item", []):
-            if service["text"] == "InteractionIds":
-                for answer in service["answer"]:
-                    data_dict["nhsAsSvcIA"].append(answer["valueString"])
-            elif service["text"] == "ManufacturingOdsCode":
-                data_dict["nhsMhsManufacturerOrg"] = service["answer"][0].get("valueString", "")
-            elif service["text"] == "PartyKey":
-                data_dict["nhsMhsPartyKey"] = service["answer"][0].get("valueString", "")
-            elif service["text"] == "Owner":
-                data_dict["nhsAsClient"] = [service["answer"][0].get("valueString", "")]
-                data_dict["nhsIdCode"] = service["answer"][0].get("valueString", "")
-    return data_dict
-
-def _raise_invalid_query_params_error():
+class EndpointCpm(BaseCpm):
+    FILTER_MAP = ENDPOINT_FILTER_MAP
+    DATA_MAP = ENDPOINT_DATA_MAP
+    DEFAULT_DICT = DEFAULT_ENDPOINT_DICT
+    
+    def __init__(self, data, query_parts):
+        self.validate_filters(query_parts)
+        super().__init__(data, query_parts)
+    
+    def validate_filters(self, query_parts):
+        non_empty_count = sum(1 for value in query_parts.values() if value and value != 0)
+        if non_empty_count < 2:
+            self._raise_invalid_query_params_error()
+    
+    def _raise_invalid_query_params_error(self):
         org_code = f'{ORG_CODE_QUERY_PARAMETER_NAME}={ORG_CODE_FHIR_IDENTIFIER}|value'
         party_key = f'{IDENTIFIER_QUERY_PARAMETER_NAME}={PARTY_KEY_FHIR_IDENTIFIER}|value'
         service_id = f'{IDENTIFIER_QUERY_PARAMETER_NAME}={SERVICE_ID_FHIR_IDENTIFIER}|value'
@@ -131,3 +149,17 @@ def _raise_invalid_query_params_error():
                         f"'{org_code}&{party_key}'"
                         f"'{service_id}&{party_key}'"
                         "]")
+
+
+class DeviceCpm(BaseCpm):
+    FILTER_MAP = DEVICE_FILTER_MAP
+    DATA_MAP = DEVICE_DATA_MAP
+    DEFAULT_DICT = DEFAULT_DEVICE_DICT
+    
+    def __init__(self, data, query_parts):
+        self.validate_filters(query_parts)
+        super().__init__(data, query_parts)
+    
+    def validate_filters(self, query_parts):
+        if "org_code" not in query_parts or "interaction_id" not in query_parts or not query_parts["org_code"] or not query_parts["interaction_id"]:
+            raise SDSException("org_code and interaction_id must be provided")