+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 2024-02-20
 - [PI-238] Filter the results returned from CPM (Device)
 - [PI-239] Translate CPM results into SDS FHIR (Device)
+- [PI-251] Translate CPM results into SDS FHIR (Endpoint)
+- [PI-252] Filter the results returned from CPM (Endpoint)
 
 ## 2024-02-09
 - [PI-233] Filter for CPM on Endpoints